refactor(build): move package entry wrappers out of src
