+++ entries/client.mjs
@@ -1,0 +1,10 @@
+// Client rendering entry. Keep this focused on the core UI primitives that are
+// useful for browser-only apps.
+
+export * as View from "../src/View.res.mjs";
+export * as Html from "../src/Html.res.mjs";
+export * as XoteJSX from "../src/XoteJSX.res.mjs";
+export * as Prop from "../src/Prop.res.mjs";
+export * as Signal from "../src/Signal.res.mjs";
+export * as Computed from "../src/Computed.res.mjs";
+export * as Effect from "../src/Effect.res.mjs";
