+++ entries/hydration.mjs
@@ -1,5 +1,5 @@
 // Hydration entry. Import explicitly in apps that hydrate server-rendered Xote
 // markup on the client.
 
-export * from "./Hydration.res.mjs";
-export * as Hydration from "./Hydration.res.mjs";
+export * from "../src/Hydration.res.mjs";
+export * as Hydration from "../src/Hydration.res.mjs";
