+++ entries/mdx.mjs
@@ -1,5 +1,5 @@
 // MDX integration entry. Kept separate from the client aggregate because MDX is
 // optional and only needed by apps compiling MDX content to Xote nodes.
 
-export * from "./Mdx.res.mjs";
-export * as Mdx from "./Mdx.res.mjs";
+export * from "../src/Mdx.res.mjs";
+export * as Mdx from "../src/Mdx.res.mjs";
