+++ entries/router.mjs
@@ -1,6 +1,6 @@
 // Router entry. Route matching is bundled with the browser router because the
 // router public API exposes route config and matching helpers together.
 
-export * from "./Router.res.mjs";
-export * as Router from "./Router.res.mjs";
-export * as Route from "./Route.res.mjs";
+export * from "../src/Router.res.mjs";
+export * as Router from "../src/Router.res.mjs";
+export * as Route from "../src/Route.res.mjs";
