+++ entries/ssr.mjs
@@ -1,0 +1,7 @@
+// Server rendering entry. SSR state and environment helpers are grouped here
+// because SSR apps usually need them alongside renderToString/renderDocument.
+
+export * from "../src/SSR.res.mjs";
+export * as SSR from "../src/SSR.res.mjs";
+export * as SSRContext from "../src/SSRContext.res.mjs";
+export * as SSRState from "../src/SSRState.res.mjs";
