--- src/client.mjs
@@ -1,10 +1,0 @@
-// Client rendering entry. Keep this focused on the core UI primitives that are
-// useful for browser-only apps.
-
-export * as View from "./View.res.mjs";
-export * as Html from "./Html.res.mjs";
-export * as XoteJSX from "./XoteJSX.res.mjs";
-export * as Prop from "./Prop.res.mjs";
-export * as Signal from "./Signal.res.mjs";
-export * as Computed from "./Computed.res.mjs";
-export * as Effect from "./Effect.res.mjs";
