--- src/index.mjs
@@ -1,5 +1,0 @@
-// Default package entry for client-rendered apps. Optional features such as
-// routing, SSR, hydration, and MDX are exposed through their own subpaths so
-// root imports do not ship the full stack by default.
-
-export * from "./client.mjs";
