--- src/ssr.mjs
@@ -1,7 +1,0 @@
-// Server rendering entry. SSR state and environment helpers are grouped here
-// because SSR apps usually need them alongside renderToString/renderDocument.
-
-export * from "./SSR.res.mjs";
-export * as SSR from "./SSR.res.mjs";
-export * as SSRContext from "./SSRContext.res.mjs";
-export * as SSRState from "./SSRState.res.mjs";
